requirements.txt
Flask~=3.0.2
Flask-SocketIO~=5.3.6
mediapipe~=0.10.2
matplotlib~=3.7.2
numpy~=1.24.3
opencv-python~=4.8.0.74
scikit-learn~=1.3.0
tensorflow~=2.16.1
google-generativeai~=0.14.0